+++ Slim/Routable.php
@@ -3,7 +3,7 @@
  * Slim Framework (https://slimframework.com)
  *
  * @link      https://github.com/slimphp/Slim
- * @copyright Copyright (c) 2011-2016 Josh Lockhart
+ * @copyright Copyright (c) 2011-2017 Josh Lockhart
  * @license   https://github.com/slimphp/Slim/blob/3.x/LICENSE.md (MIT License)
  */
 namespace Slim;
@@ -100,7 +100,7 @@ abstract class Routable
     /**
      * Set the route pattern
      *
-     * @set string
+     * @param string $newPattern
      */
     public function setPattern($newPattern)
     {