+++ tests/AppTest.php
@@ -3,8 +3,8 @@
  * Slim Framework (https://slimframework.com)
  *
  * @link      https://github.com/slimphp/Slim
- * @copyright Copyright (c) 2011-2016 Josh Lockhart
- * @license   https://github.com/slimphp/Slim/blob/master/LICENSE.md (MIT License)
+ * @copyright Copyright (c) 2011-2017 Josh Lockhart
+ * @license   https://github.com/slimphp/Slim/blob/3.x/LICENSE.md (MIT License)
  */
 
 namespace Slim\Tests;
@@ -1570,9 +1570,22 @@ class AppTest extends \PHPUnit_Framework_TestCase
     public function testRespondWithPaddedStreamFilterOutput()
     {
         $availableFilter = stream_get_filters();
-        if (in_array('mcrypt.*', $availableFilter) && in_array('mdecrypt.*', $availableFilter)) {
+
+        if (version_compare(phpversion(), '7.0.0', '>=')) {
+            $filterName           = 'string.rot13';
+            $unfilterName         = 'string.rot13';
+            $specificFilterName   = 'string.rot13';
+            $specificUnfilterName = 'string.rot13';
+        } else {
+            $filterName           = 'mcrypt.*';
+            $unfilterName         = 'mdecrypt.*';
+            $specificFilterName   = 'mcrypt.rijndael-128';
+            $specificUnfilterName = 'mdecrypt.rijndael-128';
+        }
+
+        if (in_array($filterName, $availableFilter) && in_array($unfilterName, $availableFilter)) {
             $app = new App();
-            $app->get('/foo', function ($req, $res) {
+            $app->get('/foo', function ($req, $res) use ($specificFilterName, $specificUnfilterName) {
                 $key = base64_decode('xxxxxxxxxxxxxxxx');
                 $iv = base64_decode('Z6wNDk9LogWI4HYlRu0mng==');
 
@@ -1581,7 +1594,7 @@ class AppTest extends \PHPUnit_Framework_TestCase
 
                 $stream = fopen('php://temp', 'r+');
 
-                $filter = stream_filter_append($stream, 'mcrypt.rijndael-128', STREAM_FILTER_WRITE, [
+                $filter = stream_filter_append($stream, $specificFilterName, STREAM_FILTER_WRITE, [
                     'key' => $key,
                     'iv' => $iv
                 ]);
@@ -1590,7 +1603,7 @@ class AppTest extends \PHPUnit_Framework_TestCase
                 rewind($stream);
                 stream_filter_remove($filter);
 
-                stream_filter_append($stream, 'mdecrypt.rijndael-128', STREAM_FILTER_READ, [
+                stream_filter_append($stream, $specificUnfilterName, STREAM_FILTER_READ, [
                     'key' => $key,
                     'iv' => $iv
                 ]);
@@ -1793,23 +1806,6 @@ class AppTest extends \PHPUnit_Framework_TestCase
 //        });
 //        $res = $app->run(true);
 //    }
-
-    public function testRunSlimException()
-    {
-        $app = $this->appFactory();
-        $app->get('/foo', function ($req, $res, $args) {
-            return $res;
-        });
-        $app->add(function ($req, $res, $args) {
-            $res->write("Failed");
-            throw new SlimException($req, $res);
-        });
-        $res = $app->run(true);
-
-        $res->getBody()->rewind();
-        $this->assertEquals(200, $res->getStatusCode());
-        $this->assertEquals("Failed", $res->getBody()->getContents());
-    }
 
     /**
      * @requires PHP 7.0
@@ -2064,6 +2060,32 @@ class AppTest extends \PHPUnit_Framework_TestCase
         $response = $method->invoke($app, $response);
     }
 
+    public function testUnsupportedMethodWithoutRoute()
+    {
+        $app = new App();
+        $c = $app->getContainer();
+        $c['environment'] = Environment::mock(['REQUEST_URI' => '/', 'REQUEST_METHOD' => 'BADMTHD']);
+
+        $resOut = $app->run(true);
+
+        $this->assertInstanceOf(ResponseInterface::class, $resOut);
+        $this->assertEquals(404, $resOut->getStatusCode());
+    }
+
+    public function testUnsupportedMethodWithRoute()
+    {
+        $app = new App();
+        $app->get('/', function () {
+            // stubbed action to give us a route at /
+        });
+        $c = $app->getContainer();
+        $c['environment'] = Environment::mock(['REQUEST_URI' => '/', 'REQUEST_METHOD' => 'BADMTHD']);
+
+        $resOut = $app->run(true);
+
+        $this->assertInstanceOf(ResponseInterface::class, $resOut);
+        $this->assertEquals(405, $resOut->getStatusCode());
+    }
 
     public function testContainerSetToRoute()
     {