+++ CHANGELOG.md
@@ -16,6 +16,8 @@
 - [#2067](https://github.com/slimphp/Slim/pull/2067) Remove App::VERSION
 - [#2078](https://github.com/slimphp/Slim/pull/2078) Remove App::subRequest()
 - [#2098](https://github.com/slimphp/Slim/pull/2098) Remove CallableResolverTrait
+- [#2102](https://github.com/slimphp/Slim/pull/2102) Remove container from router
+- [#2124](https://github.com/slimphp/Slim/pull/2124) Remove Slim\Exception\SlimException
 
 ### Fixed